Update: Updated to read authors, and show network status
After submission I needed to use the tags to view the authours from the Gaurdian Site, I also need to make a correction to show the staus of network connectivity.

Some articles do not have an authour, but have the author trailing the title using the "|", so I had to check if the authour was null and if the author was attached to the article title. If neither existed, "..." wiss be displayed in the author field.